requirements.txt
anthropic
playwright
pytest
pytest-playwright
allure-pytest
python-dotenv
